+++ lib/pages/camera.dart
@@ -4,7 +4,7 @@ import 'dart:typed_data';
 
 import 'package:adv_camera/adv_camera.dart';
 import 'package:adv_image_picker/adv_image_picker.dart';
-import 'package:adv_image_picker/components/toast.dart';
+import 'package:adv_image_picker/components/adv_state.dart';
 import 'package:adv_image_picker/models/result_item.dart';
 import 'package:adv_image_picker/pages/gallery.dart';
 import 'package:adv_image_picker/pages/result.dart';
@@ -13,9 +13,9 @@ import 'package:basic_components/components/adv_button.dart';
 import 'package:basic_components/components/adv_column.dart';
 import 'package:basic_components/components/adv_loading_with_barrier.dart';
 import 'package:basic_components/components/adv_visibility.dart';
+import 'package:basic_components/utilities/toast.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:path_provider/path_provider.dart';
 
 class CameraPage extends StatefulWidget {
   final bool allowMultiple;
@@ -33,10 +33,11 @@ class CameraPage extends StatefulWidget {
   }
 }
 
-void logError(String code, String message) =>
-    print('${AdvImagePicker.error}: $code\n${AdvImagePicker.errorMessage}: $message');
+void logError(String code, String message) => print(
+    '${AdvImagePicker.error}: $code\n${AdvImagePicker.errorMessage}: $message');
 
-class _CameraPageState extends State<CameraPage> with WidgetsBindingObserver {
+class _CameraPageState extends AdvState<CameraPage>
+    with WidgetsBindingObserver {
   AdvCameraController controller;
   String imagePath;
   int _currentCameraIndex = 0;
@@ -46,7 +47,7 @@ class _CameraPageState extends State<CameraPage> with WidgetsBindingObserver {
   final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
 
   @override
-  Widget build(BuildContext context) {
+  Widget buildView(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Text(
@@ -59,65 +60,74 @@ class _CameraPageState extends State<CameraPage> with WidgetsBindingObserver {
         iconTheme: IconThemeData(color: Colors.black87),
       ),
       bottomSheet: Container(
-          height: 80.0,
           padding: EdgeInsets.symmetric(horizontal: 32.0, vertical: 8.0),
           color: Colors.white,
-          child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              crossAxisAlignment: CrossAxisAlignment.end,
-              children: [
-                AdvButton.custom(
-                  child: AdvColumn(divider: ColumnDivider(4.0), children: [
+          child:
+              Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
+            AdvButton.custom(
+              child: AdvColumn(
+                  mainAxisSize: MainAxisSize.min,
+                  divider: ColumnDivider(4.0),
+                  children: [
                     Text(AdvImagePicker.rotate),
                     Icon(Icons.switch_camera),
                   ]),
-                  buttonSize: ButtonSize.small,
-                  primaryColor: Colors.white,
-                  accentColor: Colors.black87,
-                  onPressed: () {
-                    controller.switchCamera();
-                  },
-                ),
-                Container(
-                    margin: EdgeInsets.only(bottom: 8.0),
-                    child: Text(
-                      AdvImagePicker.photo,
-                      style: TextStyle(fontWeight: FontWeight.w700, fontSize: 12.0),
-                    )),
-                AdvVisibility(
-                  visibility:
-                      widget.enableGallery ? VisibilityFlag.visible : VisibilityFlag.invisible,
-                  child: AdvButton.custom(
-                    child: AdvColumn(divider: ColumnDivider(4.0), children: [
+              buttonSize: ButtonSize.small,
+              primaryColor: Colors.white,
+              accentColor: Colors.black87,
+              onPressed: () {
+                controller.switchCamera();
+              },
+            ),
+            Container(
+              margin: EdgeInsets.only(top: 8.0),
+              child: Text(
+                AdvImagePicker.photo,
+                style: TextStyle(fontWeight: FontWeight.w700, fontSize: 12.0),
+              ),
+            ),
+            AdvVisibility(
+              visibility: widget.enableGallery
+                  ? VisibilityFlag.visible
+                  : VisibilityFlag.invisible,
+              child: AdvButton.custom(
+                child: AdvColumn(
+                    mainAxisSize: MainAxisSize.min,
+                    divider: ColumnDivider(4.0),
+                    children: [
                       Text(AdvImagePicker.gallery),
                       Icon(Icons.photo_album),
                     ]),
-                    buttonSize: ButtonSize.small,
-                    primaryColor: Colors.white,
-                    accentColor: Colors.black87,
-                    onPressed: () async {
-                      if (Platform.isIOS) {
-                        bool hasPermission = await AdvImagePickerPlugin.getIosStoragePermission();
-                        if (!hasPermission) {
-                          Toast.showToast(context, "Permission denied");
-                          return null;
-                        } else {
-                          goToGallery();
-                        }
-                      } else {
-                        goToGallery();
-                      }
-                    },
-                  ),
-                ),
-              ])),
+                buttonSize: ButtonSize.small,
+                primaryColor: Colors.white,
+                accentColor: Colors.black87,
+                onPressed: () async {
+                  if (Platform.isIOS) {
+                    bool hasPermission =
+                        await AdvImagePickerPlugin.getIosStoragePermission();
+                    if (!hasPermission) {
+                      Toast.showToast(context, "Permission denied");
+                      return null;
+                    } else {
+                      goToGallery();
+                    }
+                  } else {
+                    goToGallery();
+                  }
+                },
+              ),
+            ),
+          ])),
       key: _scaffoldKey,
       body: _buildWidget(context),
       floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
       floatingActionButton: FloatingActionButton(
+        heroTag: "CaptureButton",
         elevation: 0.0,
         onPressed: () {
-          takePicture().then((resultPath) async {
+          process(() async {
+            String resultPath = await takePicture();
+
             if (resultPath == null) return;
             ByteData bytes = await _readFileByte(resultPath);
             Navigator.push(
@@ -133,7 +143,8 @@ class _CameraPageState extends State<CameraPage> with WidgetsBindingObserver {
           width: 30.0,
           height: 30.0,
           decoration: BoxDecoration(
-              color: Colors.white, borderRadius: BorderRadius.all(Radius.circular(30.0))),
+              color: Colors.white,
+              borderRadius: BorderRadius.all(Radius.circular(30.0))),
         ),
       ),
     );
@@ -147,7 +158,8 @@ class _CameraPageState extends State<CameraPage> with WidgetsBindingObserver {
       bytes = Uint8List.fromList(value);
       print('reading of bytes is completed');
     }).catchError((onError) {
-      print('Exception Error while reading audio from path:' + onError.toString());
+      print('Exception Error while reading audio from path:' +
+          onError.toString());
     });
     return bytes.buffer.asByteData();
   }
@@ -156,8 +168,7 @@ class _CameraPageState extends State<CameraPage> with WidgetsBindingObserver {
     Navigator.push(
         context,
         MaterialPageRoute(
-            builder: (BuildContext context) =>
-                GalleryPage(
+            builder: (BuildContext context) => GalleryPage(
                   allowMultiple: widget.allowMultiple,
                   maxSize: widget.maxSize,
                 )));
@@ -171,13 +182,37 @@ class _CameraPageState extends State<CameraPage> with WidgetsBindingObserver {
 
   /// Display the preview from the camera (or a message if the preview is not available).
   Widget _cameraPreviewWidget(BuildContext context) {
-    return AdvCamera(
-      onCameraCreated: _onCameraCreated,
-      onImageCaptured: (String path) {
-        takePictureCompleter.complete(path);
-        takePictureCompleter = null;
-      },
-      cameraPreviewRatio: CameraPreviewRatio.r16_9,
+    return Stack(
+      children: [
+        AdvCamera(
+          onCameraCreated: _onCameraCreated,
+          onImageCaptured: (String path) {
+            takePictureCompleter.complete(path);
+            takePictureCompleter = null;
+          },
+          cameraPreviewRatio: CameraPreviewRatio.r16_9,
+          fileNamePrefix: AdvImagePicker.cameraFilePrefixName,
+        ),
+        Positioned(
+          top: 8.0,
+          left: 8.0,
+          child: FloatingActionButton(
+            heroTag: "FlashButton",
+            backgroundColor: AdvImagePicker.primaryColor,
+            mini: true,
+            child: Icon(
+              flashType == FlashType.auto
+                  ? Icons.flash_auto
+                  : flashType == FlashType.on
+                      ? Icons.flash_on
+                      : Icons.flash_off,
+              size: 18.0,
+            ),
+            onPressed: _toggleFlash,
+            elevation: 0.0,
+          ),
+        ),
+      ],
     );
   }
 
@@ -213,14 +248,34 @@ class _CameraPageState extends State<CameraPage> with WidgetsBindingObserver {
 
   void _onCameraCreated(AdvCameraController controller) {
     this.controller = controller;
+    if (AdvImagePicker.cameraSavePath == null) {
+      AdvImagePicker.getDefaultDirectoryForCamera().then((dir) async {
+        await dir.create(recursive: true);
 
-    getApplicationDocumentsDirectory().then((Directory extDir) async {
-      final String dirPath = '${extDir.path}/Pictures';
-      await Directory(dirPath).create(recursive: true);
-
-      await controller.setSavePath(dirPath);
-    });
+        await controller
+            .setSavePath("${dir.path}/${AdvImagePicker.cameraFolderName}");
+      });
+    } else {
+      controller.setSavePath("${AdvImagePicker.cameraSavePath}");
+    }
 
     setState(() {});
+  }
+
+  void _toggleFlash() {
+    process(() async {
+      if (controller == null) return;
+
+      if (flashType == FlashType.auto) {
+        flashType = FlashType.on;
+      } else if (flashType == FlashType.on) {
+        flashType = FlashType.off;
+      } else if (flashType == FlashType.off) {
+        flashType = FlashType.auto;
+      }
+      await controller.setFlashType(flashType);
+
+      refresh();
+    });
   }
 }